+++ src/ui/components/transactions/index.tsx
@@ -20,6 +20,7 @@ import scriptInvocation from './ScriptInvocation';
 import packageTx from './Package';
 import unknown from './Unknown';
 import customData from './CustomData';
+import wavesAuth from './WavesAuth';
 
 const MESSAGES = [
     auth,
@@ -43,6 +44,7 @@ const MESSAGES = [
     scriptInvocation,
     packageTx,
     customData,
+    wavesAuth
 ];
 
 export { FinalTransaction } from './FinalTransaction';